# Conflicts:
#	variable/include/scipp/variable/transform.h

+++ variable/include/scipp/variable/structure_array_model.h
@@ -79,8 +79,10 @@ public:
     throw except::TypeError("This data type does not have bin indices.");
   }
 
-  std::span<const T> values() const { return {get_values(), size()}; }
-  std::span<T> values() { return {get_values(), size()}; }
+  std::span<const T> values() const {
+    return {get_values(), static_cast<size_t>(size())};
+  }
+  std::span<T> values() { return {get_values(), static_cast<size_t>(size())}; }
 
 private:
   const T *get_values() const;
